+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Changelog
 
 - added `mapAt(keyPath:)` operator
 - added `zip(with:)` operator
-- added `ofType()` operator
+- added `ofType(_:)` operator
 
 3.1.0
 -----